requirements.txt
# Core Framework
Flask==2.3.3
Werkzeug==2.3.7

# Authentication
Flask-Login==0.6.3
bcrypt==4.0.1

# Database
psycopg2-binary==2.9.7

# Data Processing
pandas==2.0.3
openpyxl==3.1.2

# Configuration
python-dotenv==1.0.0

# PDF and Barcode Generation
reportlab==4.0.4
python-barcode==0.15.1
Pillow==10.0.0